requirements.txt
fastapi
uvicorn
websockets
sqlalchemy
psycopg[binary]
psycopg-pool
alembic
pydantic-settings

# OpenTelemetry dependencies
opentelemetry-api
opentelemetry-sdk
opentelemetry-exporter-otlp
opentelemetry-instrumentation-fastapi
opentelemetry-instrumentation-sqlalchemy
opentelemetry-instrumentation-psycopg
opentelemetry-instrumentation-asyncio
opentelemetry-instrumentation-logging
